rename: components → ui 로 디렉토리명 변경
