+++ app/ui/bookshelf/book-thumbnail.tsx
